FundamentalFrequencies

Documentation for FundamentalFrequencies.

FundamentalFrequencies.naff

FundamentalFrequencies.naff — Function

naff(data::AbstractMatrix, n_frequencies=1; window_order=1) -> (frequencies, amplitudes)

Performs a batched Numerical Analysis of Fundamental Frequencies (NAFF) method for the (complex) signals in each row of data, computing the first n_frequencies frequencies from the signal. A Hanning window with order window_order is applied to the signal in order to gain a more accurate computation of the frequencies.

Arguments

data::AbstractMatrix: a matrix of size num_signals x n_samples to do NAFF on

n_frequencies = 1: number of frequencies to compute with NAFF

Keyword Arguments

window_order = 1: order of the Hanning window applied to the signal throughout NAFF

warnings = true: display warnings

Output

frequencies: a matrix of size num_signals x n_frequencies containing the frequencies

amplitudes: a matrix of size num_signals x n_frequencies containing the complex amplitudes associated with each frequency